Fix dashboard tab links
