+++ pages/dashboard/history.tsx
@@ -1,6 +1,4 @@
-import React from 'react';
-
-import DashboardLayout from 'sections/dashboard/DashboardContainer/DashboardLayout';
+import DashboardLayout from 'sections/dashboard/DashboardLayout';
 import History from 'sections/dashboard/History';
 
 type HistoryPageProps = React.FC & { getLayout: (page: HTMLElement) => JSX.Element };
