--- sections/dashboard/DashboardContainer/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './DashboardContainer';
